Refactor Subprocesso module to comply with ADR-002 and ADR-005
- Created `SubprocessoValidacaoController` and moved validation endpoints from `SubprocessoMapaController` (ADR-005).
- Created `EventoTransicaoSubprocesso` record and implemented event publishing in `SubprocessoTransicaoService` (ADR-002).
- Updated ADR documentation to include Status headers and reflect current architecture.
- Created `adr-report.md` summarizing the review.
- Updated unit tests to reflect controller refactoring and new service dependencies.
