+++ adr-report.md
@@ -1,0 +1,53 @@
+# Relatório de Qualidade e Conformidade de ADRs
+
+**Data:** 17 de Fevereiro de 2026
+
+## Resumo
+
+Este relatório detalha a revisão da qualidade das ADRs (Architecture Decision Records) existentes e a verificação de conformidade do código com estas decisões.
+
+## Revisão de Qualidade
+
+As seguintes ADRs foram revisadas:
+
+- **ADR-001: Facade Pattern** - Qualidade adequada.
+- **ADR-002: Unified Events** - Falta de cabeçalho de Status. Texto claro.
+- **ADR-003: Security Architecture** - Falta de cabeçalho de Status. Texto claro.
+- **ADR-004: DTO Pattern** - Falta de cabeçalho de Status. Texto claro.
+- **ADR-005: Controller Organization** - Falta de cabeçalho de Status. Texto descrevia uma organização que não estava totalmente implementada no código (faltava `SubprocessoValidacaoController`).
+- **ADR-006: Domain Aggregates** - Falta de cabeçalho de Status. Texto claro.
+- **ADR-007: Circular Dependencies** - Falta de cabeçalho de Status. Texto claro.
+- **ADR-008: Simplification Decisions** - Atua como um roadmap/changelog de arquitetura. Qualidade adequada.
+
+**Melhorias Aplicadas:**
+- Adicionado cabeçalho `Status: ✅ Ativo` em todas as ADRs que não o possuíam.
+- Padronização de formatação.
+
+## Verificação de Conformidade
+
+### Discrepâncias Identificadas e Corrigidas
+
+1.  **ADR-002 (Eventos Unificados)**
+    - **Problema:** O código utilizava chamadas diretas para notificação (`alertaService`, `emailService`) e o registro `EventoTransicaoSubprocesso` mencionado na ADR não existia.
+    - **Correção:**
+        - Criado o record `sgc.subprocesso.eventos.EventoTransicaoSubprocesso`.
+        - Atualizado `SubprocessoTransicaoService` para publicar este evento usando `ApplicationEventPublisher`.
+
+2.  **ADR-005 (Organização de Controllers)**
+    - **Problema:** A ADR descrevia a existência de `SubprocessoValidacaoController`, mas este arquivo não existia. As funcionalidades de validação estavam misturadas em `SubprocessoMapaController`.
+    - **Correção:**
+        - Criado `SubprocessoValidacaoController`.
+        - Movidos os endpoints de validação, sugestões e homologação de validação para o novo controller.
+        - Refatorado `SubprocessoMapaController` para manter apenas responsabilidades de mapa.
+
+### Conformidade Verificada (Sem Ação Necessária)
+
+- **ADR-001 (Facade Pattern):** Código segue o padrão, sem facades "pass-through" (removidas anteriormente).
+- **ADR-003 (Security Architecture):** `AccessControlService` e policies existem e são utilizados.
+- **ADR-004 (DTO Pattern):** DTOs seguem o padrão (sem formatação no backend).
+- **ADR-006 (Domain Aggregates):** Estrutura de pacotes segue o padrão.
+- **ADR-007 (Circular Dependencies):** Uso de `@Lazy` verificado em `AtividadeFacade`.
+
+## Conclusão
+
+A documentação arquitetural (ADRs) foi atualizada para refletir melhor o estado desejado e o código foi refatorado para alinhar-se com as decisões documentadas. O sistema agora está mais consistente com seus princípios arquiteturais.
