+++ backend/etc/docs/adr/ADR-004-dto-pattern.md
@@ -1,5 +1,7 @@
 # ADR-004: Padrão de DTOs Obrigatórios
 
+**Status:** ✅ Ativo
+
 ---
 
 ## Contexto
