+++ backend/etc/docs/adr/ADR-006-domain-aggregates-organization.md
@@ -1,5 +1,6 @@
 # ADR-006: Organização por Agregados de Domínio vs Outros Critérios
 
+**Status:** ✅ Ativo
 
 ---
