+++ backend/etc/docs/adr/ADR-007-circular-dependency-resolution.md
@@ -1,5 +1,7 @@
 # ADR-007: Resolução de Dependências Circulares
 
+**Status:** ✅ Ativo
+
 ---
 
 ## Contexto e Problema
