+++ backend/src/main/java/sgc/subprocesso/SubprocessoMapaController.java
@@ -11,8 +11,6 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
-import sgc.analise.AnaliseFacade;
-import sgc.analise.dto.AnaliseHistoricoDto;
 import sgc.mapa.dto.ImpactoMapaResponse;
 import sgc.mapa.dto.MapaVisualizacaoResponse;
 import sgc.mapa.dto.SalvarMapaRequest;
@@ -20,14 +18,9 @@ import sgc.mapa.model.Mapa;
 import sgc.mapa.model.MapaViews;
 import sgc.mapa.service.MapaFacade;
 import sgc.organizacao.model.Usuario;
-import sgc.comum.dto.ComumDtos.JustificativaRequest;
-import sgc.comum.dto.ComumDtos.TextoRequest;
 import sgc.subprocesso.dto.*;
 import sgc.subprocesso.model.Subprocesso;
 import sgc.subprocesso.service.SubprocessoFacade;
-
-import java.util.List;
-import java.util.Map;
 
 /**
  * Controller para operações relacionadas ao Mapa de Competências dentro do workflow de Subprocesso.
@@ -40,8 +33,6 @@ import java.util.Map;
 public class SubprocessoMapaController {
     private final SubprocessoFacade subprocessoFacade;
     private final MapaFacade mapaFacade;
-    private final AnaliseFacade analiseFacade;
-    private final sgc.comum.repo.ComumRepo repo;
 
     @GetMapping("/{codigo}/impactos-mapa")
     @PreAuthorize("isAuthenticated()")
@@ -114,92 +105,6 @@ public class SubprocessoMapaController {
 
         Mapa mapa = subprocessoFacade.salvarMapaSubprocesso(codigo, request);
         return ResponseEntity.ok(mapa);
-    }
-
-    @PostMapping("/{codigo}/apresentar-sugestoes")
-    @PreAuthorize("hasRole('CHEFE')")
-    @Operation(summary = "Apresenta sugestões de melhoria para o mapa")
-    public void apresentarSugestoes(
-            @PathVariable Long codigo,
-            @RequestBody @Valid TextoRequest request,
-            @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.apresentarSugestoes(codigo, request.texto(), usuario);
-    }
-
-    @GetMapping("/{codigo}/sugestoes")
-    @PreAuthorize("isAuthenticated()")
-    public Map<String, Object> obterSugestoes(@PathVariable Long codigo) {
-        return subprocessoFacade.obterSugestoes(codigo);
-    }
-
-    @GetMapping("/{codigo}/historico-validacao")
-    @PreAuthorize("isAuthenticated()")
-    public List<AnaliseHistoricoDto> obterHistoricoValidacao(@PathVariable Long codigo) {
-        return analiseFacade.listarHistoricoValidacao(codigo);
-    }
-
-    @PostMapping("/{codigo}/validar-mapa")
-    @PreAuthorize("hasRole('CHEFE')")
-    @Operation(summary = "Valida o mapa de competências da unidade")
-    public ResponseEntity<Void> validarMapa(@PathVariable Long codigo, @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.validarMapa(codigo, usuario);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{codigo}/devolver-validacao")
-    @PreAuthorize("hasAnyRole('ADMIN', 'GESTOR')")
-    @Operation(summary = "Devolve o mapa para ajuste (pelo chefe/gestor)")
-    public ResponseEntity<Void> devolverValidacao(
-            @PathVariable Long codigo,
-            @RequestBody @Valid JustificativaRequest request,
-            @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.devolverValidacao(codigo, request.justificativa(), usuario);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{codigo}/aceitar-validacao")
-    @PreAuthorize("hasAnyRole('ADMIN', 'GESTOR')")
-    @Operation(summary = "Aceita a validação (pelo gestor)")
-    public ResponseEntity<Void> aceitarValidacao(@PathVariable Long codigo, @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.aceitarValidacao(codigo, usuario);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{codigo}/homologar-validacao")
-    @PreAuthorize("hasRole('ADMIN')")
-    @Operation(summary = "Homologa a validação")
-    public ResponseEntity<Void> homologarValidacao(@PathVariable Long codigo, @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.homologarValidacao(codigo, usuario);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{codigo}/submeter-mapa-ajustado")
-    @PreAuthorize("hasRole('ADMIN')")
-    @Operation(summary = "Submete o mapa após ajustes solicitados")
-    public ResponseEntity<Void> submeterMapaAjustado(
-            @PathVariable Long codigo,
-            @Valid @RequestBody SubmeterMapaAjustadoRequest request,
-            @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.submeterMapaAjustado(codigo, request, usuario);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{codigo}/aceitar-validacao-bloco")
-    @PreAuthorize("hasRole('GESTOR')")
-    @Operation(summary = "Aceita validação de mapas em bloco")
-    public void aceitarValidacaoEmBloco(@PathVariable Long codigo,
-            @RequestBody @Valid ProcessarEmBlocoRequest request,
-            @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.aceitarValidacaoEmBloco(request.subprocessos(), codigo, usuario);
-    }
-
-    @PostMapping("/{codigo}/homologar-validacao-bloco")
-    @PreAuthorize("hasRole('ADMIN')")
-    @Operation(summary = "Homologa validação de mapas em bloco")
-    public void homologarValidacaoEmBloco(@PathVariable Long codigo,
-            @RequestBody @Valid ProcessarEmBlocoRequest request,
-            @AuthenticationPrincipal Usuario usuario) {
-        subprocessoFacade.homologarValidacaoEmBloco(request.subprocessos(), codigo, usuario);
     }
 
     @PostMapping("/{codigo}/disponibilizar-mapa-bloco")
