+++ backend/src/test/java/sgc/subprocesso/SubprocessoMapaControllerTest.java
@@ -8,22 +8,18 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.ResponseEntity;
-import sgc.analise.AnaliseFacade;
-import sgc.analise.dto.AnaliseHistoricoDto;
 import sgc.mapa.dto.ImpactoMapaResponse;
 import sgc.mapa.dto.MapaVisualizacaoResponse;
 import sgc.mapa.dto.SalvarMapaRequest;
 import sgc.mapa.model.Mapa;
 import sgc.mapa.service.MapaFacade;
 import sgc.organizacao.model.Usuario;
-import sgc.comum.dto.ComumDtos.*;
 import sgc.subprocesso.dto.*;
 import sgc.subprocesso.model.Subprocesso;
 import sgc.subprocesso.service.SubprocessoFacade;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -44,9 +40,6 @@ class SubprocessoMapaControllerTest {
 
     @Mock
     private MapaFacade mapaFacade;
-
-    @Mock
-    private AnaliseFacade analiseFacade;
 
     @Test
     @DisplayName("Deve verificar impactos")
@@ -147,129 +140,6 @@ class SubprocessoMapaControllerTest {
         ResponseEntity<Mapa> response = controller.salvarMapaCompleto(codigo, req);
 
         assertThat(response.getBody()).isEqualTo(mapa);
-    }
-
-    @Test
-    @DisplayName("Deve apresentar sugestões")
-    void deveApresentarSugestoes() {
-        Long codigo = 1L;
-        TextoRequest req = new TextoRequest("Sugestão");
-        Usuario usuario = new Usuario();
-
-        controller.apresentarSugestoes(codigo, req, usuario);
-
-        verify(subprocessoFacade).apresentarSugestoes(codigo, "Sugestão", usuario);
-    }
-
-    @Test
-    @DisplayName("Deve obter sugestões")
-    void deveObterSugestoes() {
-        Long codigo = 1L;
-        Map<String, Object> dto = Map.of("sugestoes", "Texto");
-        when(subprocessoFacade.obterSugestoes(codigo)).thenReturn(dto);
-
-        Map<String, Object> result = controller.obterSugestoes(codigo);
-
-        assertThat(result).isEqualTo(dto);
-    }
-
-    @Test
-    @DisplayName("Deve obter histórico de validação")
-    void deveObterHistoricoValidacao() {
-        Long codigo = 1L;
-        when(analiseFacade.listarHistoricoValidacao(codigo)).thenReturn(List.of());
-
-        List<AnaliseHistoricoDto> result = controller.obterHistoricoValidacao(codigo);
-
-        assertThat(result).isEmpty();
-    }
-
-    @Test
-    @DisplayName("Deve validar mapa")
-    void deveValidarMapa() {
-        Long codigo = 1L;
-        Usuario usuario = new Usuario();
-
-        ResponseEntity<Void> response = controller.validarMapa(codigo, usuario);
-
-        verify(subprocessoFacade).validarMapa(codigo, usuario);
-        assertThat(response.getStatusCode().value()).isEqualTo(200);
-    }
-
-    @Test
-    @DisplayName("Deve devolver validação")
-    void deveDevolverValidacao() {
-        Long codigo = 1L;
-        JustificativaRequest req = new JustificativaRequest("Justificativa");
-        Usuario usuario = new Usuario();
-
-        ResponseEntity<Void> response = controller.devolverValidacao(codigo, req, usuario);
-
-        verify(subprocessoFacade).devolverValidacao(codigo, "Justificativa", usuario);
-        assertThat(response.getStatusCode().value()).isEqualTo(200);
-    }
-
-    @Test
-    @DisplayName("Deve aceitar validação")
-    void deveAceitarValidacao() {
-        Long codigo = 1L;
-        Usuario usuario = new Usuario();
-
-        ResponseEntity<Void> response = controller.aceitarValidacao(codigo, usuario);
-
-        verify(subprocessoFacade).aceitarValidacao(codigo, usuario);
-        assertThat(response.getStatusCode().value()).isEqualTo(200);
-    }
-
-    @Test
-    @DisplayName("Deve homologar validação")
-    void deveHomologarValidacao() {
-        Long codigo = 1L;
-        Usuario usuario = new Usuario();
-
-        ResponseEntity<Void> response = controller.homologarValidacao(codigo, usuario);
-
-        verify(subprocessoFacade).homologarValidacao(codigo, usuario);
-        assertThat(response.getStatusCode().value()).isEqualTo(200);
-    }
-
-    @Test
-    @DisplayName("Deve submeter mapa ajustado")
-    void deveSubmeterMapaAjustado() {
-        Long codigo = 1L;
-        SubmeterMapaAjustadoRequest req = SubmeterMapaAjustadoRequest.builder().build();
-        Usuario usuario = new Usuario();
-
-        ResponseEntity<Void> response = controller.submeterMapaAjustado(codigo, req, usuario);
-
-        verify(subprocessoFacade).submeterMapaAjustado(codigo, req, usuario);
-        assertThat(response.getStatusCode().value()).isEqualTo(200);
-    }
-
-    @Test
-    @DisplayName("Deve aceitar validação em bloco")
-    void deveAceitarValidacaoEmBloco() {
-        Long codigo = 100L;
-        List<Long> subprocessos = List.of(1L, 2L);
-        ProcessarEmBlocoRequest req = new ProcessarEmBlocoRequest("ACEITAR", subprocessos, null);
-        Usuario usuario = new Usuario();
-
-        controller.aceitarValidacaoEmBloco(codigo, req, usuario);
-
-        verify(subprocessoFacade).aceitarValidacaoEmBloco(subprocessos, codigo, usuario);
-    }
-
-    @Test
-    @DisplayName("Deve homologar validação em bloco")
-    void deveHomologarValidacaoEmBloco() {
-        Long codigo = 100L;
-        List<Long> subprocessos = List.of(1L, 2L);
-        ProcessarEmBlocoRequest req = new ProcessarEmBlocoRequest("HOMOLOGAR", subprocessos, null);
-        Usuario usuario = new Usuario();
-
-        controller.homologarValidacaoEmBloco(codigo, req, usuario);
-
-        verify(subprocessoFacade).homologarValidacaoEmBloco(subprocessos, codigo, usuario);
     }
 
     @Test
